Add upvotes functionallity -> client & server

+++ Product-Feedback-App/backend/routes/feedbacks.js
@@ -7,9 +7,10 @@ const router = express.Router();
 router.post('', checkAuth, async (req, res, next) => {
     
     const feedback = new Feedback({
+        author: req.userData.id,
         title: req.body.title,
         category: req.body.category,
-        upvotes: req.body.upvotes,
+        upvotes: [],
         status: req.body.status,
         description: req.body.description
     });
@@ -24,39 +25,69 @@ router.post('', checkAuth, async (req, res, next) => {
 router.delete('/:id', checkAuth, async (req, res, next) => {
     
     try {
-        await Feedback.deleteOne({ _id: req.params.id });
+        await Feedback.deleteOne({ _id: req.params.id, author: req.userData.id });
         res.status(200).json({
             message: 'Post deleted!'
         });
     } catch {
-        //
+        res.status(401).json({
+            message: 'Wrong author of the feedback!'
+        })
     }
 })
 
 router.patch('/:id', checkAuth, async (req, res, next) => {
     try {
         const feedback = await Feedback.findById({ _id: req.params.id });
-        if (!feedback) {
-            res.status(404).json({
-                message: 'Feedback not found!'
+
+        if (feedback.author == req.userData.id) { 
+            feedback.title = req.body.title;
+            feedback.category = req.body.category;
+            feedback.status = req.body.status;
+            feedback.description = req.body.description;
+            await feedback.save();
+    
+            res.status(200).json({
+                message: 'Update succesfull!'
             })
+        } else {
+            res.status(401).json({
+                message: 'Wrong author of the feedback!'
+            }) 
         }
-
-        feedback.title = req.body.title;
-        feedback.category = req.body.category;
-        feedback.upvotes = req.body.upvotes;
-        feedback.status = req.body.status;
-        feedback.description = req.body.description;
-        await feedback.save();
-
-        res.status(200).json({
-            message: 'Update succesfull!'
-        })
-
     } catch {
         res.status(500).json({
             message: 'An error occurred!'
         });
+    }
+})
+
+router.patch('/upvotes/:id', checkAuth, async (req, res, next) => {
+    try {
+        const feedback = await Feedback.findById({ _id: req.params.id });
+        const userID = req.userData.id;
+        const idx = feedback.upvotes.indexOf(userID);
+
+        console.log(feedback);
+        console.log(userID);
+        console.log(idx);
+
+        if (idx === -1) { 
+            feedback.upvotes.push(userID);
+        } else {
+            feedback.upvotes.splice(idx, 1);
+        }
+
+        await feedback.save();
+
+        res.status(200).json({
+            message: 'Upvote updated successfully'
+        });
+
+    } catch {
+        res.status(500).json({
+            message: 'En error occured!'
+        })
     }
 })
